Move most of Parameter functionality down to Swagger2BaseParameter, since OpenAPI3 is fairly different
